% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeGenerators.R
\name{markerPureDom}
\alias{markerPureDom}
\alias{markerPureRec}
\alias{markerHybrid}
\alias{markerRandom}
\title{Simulate genetic encodings}
\usage{
markerPureDom(n)

markerPureRec(n)

markerHybrid(n)

markerRandom(n)
}
\arguments{
\item{n}{integer number of markers}
}
\value{
A 2-column numeric matrix of encodings.
}
\description{
These functions simulate the given number of marker
encodings under different settings.
}
\details{
Encodings

These functions all accept `n`, the number of markers,
and generate a two column matrix of encodings. The included
functions all correspond to known crosses or settings useful to
experiment with recombination.
}
\section{Functions}{
\itemize{
\item \code{markerPureDom()}: Simulating a pure dominant encoding

\item \code{markerPureRec()}: Simulating a pure recessive encoding

\item \code{markerHybrid()}: Simulating a hybrid encoding

\item \code{markerRandom()}: Simulating a random encoding

}}
\author{
Chris Salahub
}
