% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{filterPopulation}
\alias{filterPopulation}
\title{Remove encodings from a `population` based on a rule}
\usage{
filterPopulation(population, rule = function(mat) any(is.na(mat)))
}
\arguments{
\item{population}{`population` with encodings to be filtered}

\item{rule}{function to apply to each encoding that returns a
boolean, defaults to checking if any values are NA}
}
\value{
A population with only those encodings for which rule
returns `TRUE`
}
\description{
Remove encodings from a `population` based on a rule
}
\details{
A simple wrapper function which applies a function that
accepts an encoding matrix and returns a logical to every encoding
in a `population` and keeps only those which return `TRUE`
}
\author{
Chris Salahub
}
