% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeGenerators.R
\name{generateBiAllelic}
\alias{generateBiAllelic}
\title{Generate allele settings}
\usage{
generateBiAllelic(nmark)
}
\arguments{
\item{nmark}{named integer vector of the count of measured markers
by chromosome}
}
\value{
A list the same length as the sum of nmark with each entry
`c("A", "a")`
}
\description{
Generate allele settings
}
\details{
This function, called when alleles are not specified in
`makeGenome`, creates a list by repeating `c("A", "a")`.
}
\author{
Chris Salahub
}
