% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeGenerators.R
\name{generateEncoding}
\alias{generateEncoding}
\title{Generate a genetic encoding}
\usage{
generateEncoding(nmark, markerFuns = markerPureDom, ...)
}
\arguments{
\item{nmark}{named integer vector of measured markers by
chromosome}

\item{markerFuns}{list of functions which each generate encodings
for a chromosome given `n`, the number of markers on the
chromosome}

\item{...}{list of additional arguments to each function in
`markerFuns`, recycled if not the same length}
}
\value{
A two-column matrix of encodings for the entire genome.
}
\description{
Generate a genetic encoding
}
\details{
A wrapper function for marker-generating functions,
this takes a list of marker-generating functions and an integer
vector of counts of the same length and generates encodings by
calling each marker function on the corresponding count. The
marker functions are recycled by default.
}
\author{
Chris Salahub
}
