% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeGenerators.R
\name{locationUniform}
\alias{locationUniform}
\alias{locationRegular}
\title{Simulate marker locations}
\usage{
locationUniform(nmark, min = 0, max = 100)

locationRegular(nmark, delta = 50)
}
\arguments{
\item{nmark}{integer number of markers}

\item{min}{smallest numeric location, must be zero or greater}

\item{max}{largest numeric location}

\item{delta}{constant distance to repeat}
}
\value{
A vector of length `nmark` giving marker locations.
}
\description{
These functions generate locations given a number of
markers based on different settings.
}
\details{
Locations

These functions accept an integer `n` and potentially
other arguments and return a vector of length `n` giving locations
on a chromosome. The default included functions are
`locationUniform`, which randomly samples uniform locations
between `min` and `max` cM, and `locationRegular`, which assumes
a constant distance of `delta` cM between each marker.
}
\section{Functions}{
\itemize{
\item \code{locationUniform()}: Simulating uniform locations

\item \code{locationRegular()}: Simulating evenly-spaced locations

}}
\author{
Chris Salahub
}
