% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomes.R
\name{makeGenome}
\alias{makeGenome}
\title{Making a genome from provided slots}
\usage{
makeGenome(
  location,
  alleles,
  chromosome,
  encoding = markerHybrid(length(chromosome))
)
}
\arguments{
\item{location}{list of marker locations by chromosome}

\item{alleles}{list of alleles by marker location}

\item{chromosome}{factor giving chromosome values with levels
ordered as they should appear in plots}

\item{encoding}{numeric two-column matrix with the same number
of rows as the length of `chromosome`, `alleles`}
}
\value{
A genome object with slots given by the provided
arguments.
}
\description{
`makeGenome` checks arguments for conformity and
then stores them in a `genome` object. `genome` is an S3 class
representing the encoding matrix of a diploid genome.
}
\details{
A simple constructor that checks arguments and provides
descriptive errors to ensure a valid genome is constructed.
}
\author{
Chris Salahub
}
