% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{selectGenome}
\alias{selectGenome}
\title{Selecting a genome from a population}
\usage{
selectGenome(population, ind = 1)
}
\arguments{
\item{population}{`population` to be subset by markers}

\item{ind}{index of individual encoding to extract}
}
\value{
A `genome` with the encoding of the individual at
`ind` in `population`.
}
\description{
Selecting a genome from a population
}
\details{
This function takes an individual encoding from a
population and converts it back to a `genome` object for
plotting and manipulation.
}
\author{
Chris Salahub
}
