% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{theoryCorrelation}
\alias{theoryCorrelation}
\title{Compute the theoretical correlation in a genome}
\usage{
theoryCorrelation(genome, map = mapHaldane, setting = "intercross")
}
\arguments{
\item{genome}{`genome` or `population` with slots giving
chromosomes and locations of markers}

\item{map}{function which accepts a vector of numerics giving
distances between locations and returns a vector of probabilities
of recombination for the corresponding distances}

\item{setting}{population setting (one of backcross, intercross,
or halfback) that determines the constant for computing
correlation}
}
\value{
The correlation derived from the distances between markers
in `genome` of `population`.
}
\description{
Compute the theoretical correlation in a genome
}
\details{
Using a given map distance function and population cross
setting, returns the theoretical correlation, proportional to
$$1/2 - p_r(d)/2,$$
where $p_r(d)$ is the probability of recombination for markers
with distance $d$ between them.
}
\author{
Chris Salahub
}
